Show more content on bridge pages
