+++ packages/frontend/src/pages/bridges-projects/view/ProjectDetails.tsx
@@ -1,6 +1,10 @@
 import React from 'react'
 
 import {
+  ContractsSection,
+  ContractsSectionProps,
+} from '../../../components/project/ContractsSection'
+import {
   DescriptionSection,
   DescriptionSectionProps,
 } from '../../../components/project/DescriptionSection'
@@ -12,11 +16,27 @@ import {
   NewsSection,
   NewsSectionProps,
 } from '../../../components/project/NewsSection'
+import {
+  PermissionsSection,
+  PermissionsSectionProps,
+} from '../../../components/project/PermissionsSection'
+import {
+  TechnologyIncomplete,
+  TechnologyIncompleteProps,
+} from '../../../components/project/TechnologyIncomplete'
+import {
+  TechnologySection,
+  TechnologySectionProps,
+} from '../../../components/project/TechnologySection'
 
 export interface ProjectDetailsProps {
   linkSection: LinkSectionProps
   newsSection: NewsSectionProps
   descriptionSection: DescriptionSectionProps
+  incomplete?: TechnologyIncompleteProps
+  sections: TechnologySectionProps[]
+  permissionsSection?: PermissionsSectionProps
+  contractsSection: ContractsSectionProps
 }
 
 export function ProjectDetails(props: ProjectDetailsProps) {
@@ -26,6 +46,15 @@ export function ProjectDetails(props: ProjectDetailsProps) {
       <NewsSection {...props.newsSection} />
       <div className="ProjectDetails-Content">
         <DescriptionSection {...props.descriptionSection} />
+        {/* <RiskSection {...props.riskSection} /> */}
+        {props.incomplete && <TechnologyIncomplete {...props.incomplete} />}
+        {props.sections.map((section) => (
+          <TechnologySection key={section.id} {...section} />
+        ))}
+        {props.permissionsSection && (
+          <PermissionsSection {...props.permissionsSection} />
+        )}
+        <ContractsSection {...props.contractsSection} />
       </div>
     </main>
   )
